% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCCRstacking.R
\name{BCCRstacking}
\alias{BCCRstacking}
\title{An Efficient Genome Selection Program for Big Data Based on Machine Learning}
\usage{
BCCRstacking(
  y.train = y.train,
  X.train = X.train,
  y.test = NULL,
  X.test = NULL,
  CV = FALSE,
  cc_num = NULL
)
}
\arguments{
\item{y.train}{a phenotype vector of type numeric}

\item{X.train}{a genotype matrix of type numeric}

\item{y.test}{a phenotype vector of type numeric, the phenotype of verification group}

\item{X.test}{a genotype matrix of type numeric, the verification group(or no phenotype group) required to calculate GEBV}

\item{CV}{a logical variable, decide whether to include covariates when compressing components}

\item{cc_num}{the number of compressed components}
}
\value{
a list including compressed components and GEBV(need provide X.test)
}
\description{
\code{BCCRstacking}calculation of GEBV based on gradient boosting compressed component regression algorithm and stacking ensemble learning algorithm
}
\examples{
data(example_dataset)
library(Rfast)
y.train <- phenotype[c(1:200),1]
X.train <- genotype[c(1:200),]
X.test <- genotype[-c(1:200),]
BCCRS <- BCCRstacking(y.train = y.train, X.train = X.train, X.test = X.test, CV=FALSE, cc_num=NULL)

}
